Renamed wasmer_wasm.h to wasmer.h
