+++ lib/c-api/README.md
@@ -14,7 +14,7 @@ $WASMER_DIR/
     wasm.h
     wasmer.h
     wasmer.hh
-    wasmer_wasm.h
+    wasmer.h
 ```
 
 Wasmer binary also ships with [`wasmer-config`](#wasmer-config)
@@ -27,7 +27,7 @@ Here is a simple example to use the C API:
 
 ```c
 #include <stdio.h>
-#include "wasmer_wasm.h"
+#include "wasmer.h"
 
 int main(int argc, const char* argv[]) {
     const char *wat_string =
