+++ lib/c-api/examples/Makefile
@@ -18,7 +18,13 @@ $(info * CFLAGS: $(CFLAGS))
 $(info * LDFLAGS: $(LDFLAGS))
 $(info * LDLIBS: $(LDLIBS))
 
-ALL = instance imports-exports exports-function exports-global memory features
+ALL = deprecated-header early-exit instance imports-exports exports-function exports-global memory memory2 features wasi
+
+.SILENT: deprecated-header deprecated-header.o
+deprecated-header: deprecated-header.o
+
+.SILENT: early-exit early-exit.o
+early-exit: early-exit.o
 
 .SILENT: instance instance.o
 instance: instance.o
@@ -35,8 +41,14 @@ exports-global: exports-global.o
 .SILENT: memory memory.o
 memory: memory.o
 
+.SILENT: memory2 memory2.o
+memory2: memory2.o
+
 .SILENT: features features.o
 features: features.o
+
+.SILENT: wasi wasi.o
+wasi: wasi.o
 
 .PHONY: all
 all: $(ALL)
