+++ lib/c-api/examples/assets/call_trap.wat
