+++ lib/c-api/examples/assets/qjs.wasm
