+++ lib/c-api/src/wasm_c_api/instance.rs
@@ -99,7 +99,7 @@ pub unsafe extern "C" fn wasm_instance_delete(_instance: Option<Box<wasm_instanc
 /// # use inline_c::assert_c;
 /// # fn main() {
 /// #    (assert_c! {
-/// # #include "tests/wasmer_wasm.h"
+/// # #include "tests/wasmer.h"
 /// #
 /// int main() {
 ///     // Create the engine and the store.
@@ -201,7 +201,7 @@ mod tests {
     #[test]
     fn test_instance_new() {
         (assert_c! {
-            #include "tests/wasmer_wasm.h"
+            #include "tests/wasmer.h"
 
             // The `sum` host function implementation.
             wasm_trap_t* sum_callback(
