+++ lib/c-api/src/wasm_c_api/unstable/features.rs
@@ -8,7 +8,7 @@
 //! # use inline_c::assert_c;
 //! # fn main() {
 //! #    (assert_c! {
-//! # #include "tests/wasmer_wasm.h"
+//! # #include "tests/wasmer.h"
 //! #
 //! int main() {
 //!     // Declare features.
