+++ lib/c-api/src/wasm_c_api/value.rs
@@ -46,7 +46,7 @@ pub union wasm_val_inner {
 /// # use inline_c::assert_c;
 /// # fn main() {
 /// #    (assert_c! {
-/// # #include "tests/wasmer_wasm.h"
+/// # #include "tests/wasmer.h"
 /// #
 /// int main() {
 ///     // Create a 32-bit integer Wasm value.
