+++ lib/cli/src/commands/wasmer_create_exe_main.c
@@ -1,4 +1,4 @@
-#include "wasmer_wasm.h"
+#include "wasmer.h"
 #include "my_wasm.h"
 
 #include <stdio.h>
