+++ lib/engine-staticlib/README.md
@@ -29,7 +29,7 @@ Target: x86_64-apple-darwin
 Now let's create a program to link with this static object file.
 
 ```c
-#include "wasmer_wasm.h"
+#include "wasmer.h"
 #include "my_wasm.h"
 
 #include <stdio.h>
