+++ tests/integration/cli/tests/staticlib_engine_test_c_source.c
@@ -1,4 +1,4 @@
-#include "wasmer_wasm.h"
+#include "wasmer.h"
 #include "my_wasm.h"
 
 #include <stdio.h>
